src/externals/GazeMTL_pytorch/requirements.txt
# Modern PyTorch-based requirements for GazeMTL
# Latest stable versions as of 2024

# Core PyTorch ecosystem
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Metrics and evaluation
torchmetrics>=1.2.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
Pillow>=10.0.0
pydicom>=2.4.0

# Visualization and logging (optional but recommended)
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0

# Optional: For advanced features
# wandb>=0.15.0  # Uncomment if you want to use Weights & Biases
# pytorch-lightning>=2.1.0  # Uncomment if you want to use PyTorch Lightning

